Refactor triangulate/tetrahedralize
